Customizing the site for PiWeatherRock
This commit includes all the real setup of this site and the migration
of documentation from README.md in PiWeatherRock's repository.

The image used on the front page was taken by me so I don't have to
worry any about copyright or anything.

This commit also includes a lot of deletes. This is simply a byproduct
of taking a them that came with sample content and making it my own.